requirements.txt
app==0.0.1
bidict==0.22.1
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
gunicorn==21.2.0
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
python-dotenv==1.0.0
python-engineio==4.8.2
python-socketio==5.11.0
setuptools==69.0.3
simple-websocket==1.0.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wheel==0.42.0
wsproto==1.2.0
